Let the server name be specified just once.
If only one of '--server-name' or '-N' (chef node name) is given, use
that name for both attributes, and only fall back to the server ID for
the Chef node name if there's no names given at all.

This makes it much simpler to keep the same name on Brightbox and in
Chef, especially if you use autogenerated names (you can just use
backticks and call your autogeneration program in the command-line).