Guard against calling `#name` on nil resources
In a few edge cases, an associated resource may no longer be directly
accessible, which will result in `nil` being returned.

When an image was public, used to build a server and then set to private
then the image's full details are not available to the API when
chaining.

So `server.image` returns `nil` whilst `server#image_id` returns
`img-12345` since the identifier is data on the server object.

This resolves #21